+++ extra_platforms/platform_data.py
@@ -44,6 +44,21 @@ from .trait import Platform
 
 AIX = Platform("aix", "IBM AIX", "➿", "https://ibm.com/products/aix/")
 
+ALMALINUX = Platform("almalinux", "AlmaLinux", "", "https://almalinux.org")
+"""
+```{note}
+AlmaLinux has no representative Unicode emoji, so like
+{data}`~extra_platforms.NOBARA` it uses a [NerdFont](https://www.nerdfonts.com)
+icon instead:
+[`` (i.e. `nf-linux-almalinux` /
+`f31d`)](https://www.nerdfonts.com/cheat-sheet?glyphSearch=nf-linux-almalinux).
+
+The side-effect of using a NerdFont character is it will only display correctly
+when using a supported font. Otherwise, it will appear as an unknown or invisible
+character depending on the fonts.
+```
+"""
+
 ALPINE = Platform("alpine", "Alpine Linux", "🏔️", "https://alpinelinux.org")
 
 ALTLINUX = Platform("altlinux", "ALT Linux", "Δ", "https://altlinux.org")